+++ django_spicy_id/tests/fields_test.py
@@ -1,6 +1,7 @@
 from unittest import mock
 
 from django.core.exceptions import ImproperlyConfigured
+from django.db.utils import ProgrammingError
 from django.test import TestCase
 
 from django_spicy_id.errors import MalformedSpicyIdError
@@ -57,7 +58,7 @@ class TestFields(TestCase):
 
         # When padding is disabled, it's an error to use padding characters.
         self.assertTrue(model.objects.filter(id="ex_8M0kX").first())
-        with self.assertRaises(MalformedSpicyIdError):
+        with self.assertRaises(ProgrammingError):
             model.objects.filter(id="ex_0008M0kX").first()
 
         boundary = model.objects.create(id=2**63 - 1)
@@ -80,7 +81,7 @@ class TestFields(TestCase):
 
         # Using uppercase hex characters (i.e. supporting multiple legal
         # representations of the same value) is not allowed.
-        with self.assertRaises(MalformedSpicyIdError):
+        with self.assertRaises(ProgrammingError):
             model.objects.filter(id="ex_75BCD15").first()
 
         boundary = model.objects.create(id=2**63 - 1)
@@ -151,7 +152,7 @@ class TestFields(TestCase):
         self.assertEqual(retrieved, o)
 
         # Exact padding characters are mandatory when configured on the field.
-        with self.assertRaises(MalformedSpicyIdError):
+        with self.assertRaises(ProgrammingError):
             model.objects.filter(pk="ex_0008M0kX").first()
         self.assertEqual(retrieved, o)
 
@@ -165,22 +166,22 @@ class TestFields(TestCase):
         self.assertEqual(retrieved, o)
 
         # Exact padding characters are mandatory when configured on the field.
-        with self.assertRaises(MalformedSpicyIdError):
+        with self.assertRaises(ProgrammingError):
             model.objects.filter(pk="ex_0075bcd15").first()
         self.assertEqual(retrieved, o)
 
     def test_base62_model_create_by_string(self):
         model = models.Base62Model_WithPadding
-        o = model.objects.create(id="ex_7j")
+        o = model.objects.create(id="ex_0000000007j")
         self.assertEqual("ex_0000000007j", o.id)
 
-    def test_base62_model_create_by_string(self):
+    def test_hex_model_create_by_string(self):
         model = models.HexModel_WithPadding
         o = model.objects.create(id="ex_0000000000000123")
         self.assertEqual("ex_0000000000000123", o.id)
 
         # Exact padding characters are mandatory when configured on the field.
-        with self.assertRaises(MalformedSpicyIdError):
+        with self.assertRaises(ProgrammingError):
             model.objects.create(id="ex_000124")
 
     def test_base62_model_create_by_integer(self):